feat: refactor package structure and implement ViewModels for Main and Detail screens
