+++ app/src/main/java/uvg/edu/lab6/data/model/Pokemon.kt
