+++ app/src/main/java/uvg/edu/lab6/data/remote/ApiService.kt.kt
@@ -1,12 +1,12 @@
-package uvg.edu.lab6.data.network
+package uvg.edu.lab6.data.remote
 
-import uvg.edu.lab6.data.model.Pokemon
-import uvg.edu.lab6.data.model.PokemonListResponse
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
+import uvg.edu.lab6.data.model.Pokemon
+import uvg.edu.lab6.data.model.PokemonListResponse
 
-interface PokeApiService {
+interface ApiService {
     @GET("pokemon")
     suspend fun getPokemonList(@Query("limit") limit: Int = 100): PokemonListResponse
