+++ app/src/main/java/uvg/edu/lab6/data/remote/RetrofitClient.kt.kt
@@ -1,16 +1,16 @@
-package uvg.edu.lab6.data.network
+package uvg.edu.lab6.data.remote
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-object RetrofitInstance {
+object RetrofitClient {
     private const val BASE_URL = "https://pokeapi.co/api/v2/"
 
-    val api: PokeApiService by lazy {
+    val api: ApiService by lazy {
         Retrofit.Builder()
             .baseUrl(BASE_URL)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
-            .create(PokeApiService::class.java)
+            .create(ApiService::class.java)
     }
 }
