+++ app/src/main/java/uvg/edu/lab6/data/repository/MainRepository.kt
@@ -2,10 +2,9 @@ package uvg.edu.lab6.data.repository
 
 import uvg.edu.lab6.data.model.Pokemon
 import uvg.edu.lab6.data.model.PokemonListResponse
-import uvg.edu.lab6.data.network.RetrofitInstance
 
-class PokemonRepository {
-    private val api = RetrofitInstance.api
+class MainRepository {
+    private val api = `RetrofitClient.kt`.api
 
     suspend fun getPokemonList(limit: Int = 150): PokemonListResponse {
         return api.getPokemonList(limit)
