--- app/src/main/java/uvg/edu/lab6/main/PokemonListItem.kt
@@ -1,4 +1,0 @@
-package uvg.edu.lab6.main
-
-class PokemonListItem {
-}
