+++ app/src/main/java/uvg/edu/lab6/ui/main/MainFragment.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
 import uvg.edu.lab6.data.model.PokemonSummary
+import uvg.edu.lab6.ui.MainViewModel
 import uvg.edu.lab6.ui.theme.PokedexTheme
 
 @Composable
