+++ app/src/main/java/uvg/edu/lab6/ui/main/PokemonListItem.kt
@@ -1,0 +1,4 @@
+package uvg.edu.lab6.ui.main
+
+class PokemonListItem {
+}
